Integration nextapi

+++ src/Classes/Models/onlinevideomodel.cpp
@@ -20,7 +20,6 @@
 #include <QJsonObject>
 #include <QJsonArray>
 #include "onlinevideomodel.h"
-#include "../../globalconstants.h"
 
 OnlineVideoModel::OnlineVideoModel()
 {
@@ -91,7 +90,7 @@ void OnlineVideoModel::writeToJson(QJsonObject &json) const noexcept
     json["skips"] = skipsObject;
 }
 
-void OnlineVideoModel::setId(const int id) noexcept
+void OnlineVideoModel::setId(const double id) noexcept
 {
     m_Id = id;
 }
